update: UI/UX improvements(profile, gallery- backend)

+++ backend/models/place.py
@@ -1,5 +1,5 @@
 #!/usr/bin/env python3
-from sqlalchemy import Column, String, Integer, DateTime, func, Float
+from sqlalchemy import Column, String, Integer, DateTime, func, Float, DECIMAL
 from sqlalchemy.orm import relationship
 from .base_model import Base
 
@@ -11,8 +11,8 @@ class Place(Base):
     country = Column(String(150), nullable=False, index=True)
     region = Column(String(100), nullable=False, index=True)
     sub_region = Column(String(100), nullable=False)
-    longitude = Column(Float, nullable=True)
-    latitude = Column(Float, nullable=True)
+    longitude = Column(DECIMAL(11, 8), nullable=True)
+    latitude = Column(DECIMAL(10, 8), nullable=True)
 
     user = relationship("User", uselist=False, back_populates="place")
